Now both versions working
